requirements.txt
flask==3.0.3
flask-restx==1.3.0
pymongo==4.8.0
python-dotenv==1.0.1
